made models work and deleted unnnesecasay things in services

+++ demo/src/main/java/com/example/demo/controller/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
         this.userService = userService;
     }
     @GetMapping("/{id}")
-    public ResponseEntity<UserDto> getUserById(@PathVariable Integer id) {
+    public ResponseEntity<UserDto> getUserById(@PathVariable Long id) {
         try {
             return ResponseEntity.ok(userService.getUserById(id));
         } catch (IllegalArgumentException e) {
