+++ demo/src/main/java/com/example/demo/dto/UserDto.java
@@ -1,24 +1,24 @@
 package com.example.demo.dto;
 
 public class UserDto {
-    private int id;
+    private Long id;
     private String username;
     private String password;
 
     public UserDto() {}
 
-    public UserDto(int id, String username,String password) {
+    public UserDto(Long id, String username,String password) {
         this.id = id;
         this.username = username;
         this.password = password;
 
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
