+++ demo/src/main/java/com/example/demo/model/Votable.java
@@ -2,5 +2,5 @@ package com.example.demo.model;
 
 
 public interface Votable {
-    int getId();  // poți pune metode utile pentru identificare
+    Long   getId();  // poți pune metode utile pentru identificare
 }
